Add debug print to nvcc_wrapper when env var NVCC_WRAPPER_SHOW_COMMANDS_BEING_RUN=1 (#19)
This allows you to see the actually nvcc commands that are run by the
nvcc_wrapper while it is running.  This is very useful for debugging problems
with the nvcc_wrapper script and problem with nvcc itself.

I needed this to help debug issues with resoruce file handling with Ninja for
trilinos/Trilinos#3069.